Move get_following_status implementation to new file
